# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Minizon API",
        "description": "API REST de la plateforme Minizon — Transport et livraison.",
        "contact": {
            "name": "Équipe Minizon",
            "email": "dev@minizon.com"
        },
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000",
            "description": "Serveur local"
        }
    ],
    "paths": {
...
        "/api/withdrawals": {
            "get": {
                "tags": [
                    "💸 Retraits & Virements"
                ],
                "summary": "Mes retraits (conducteur)",
                "operationId": "withdrawalIndex",
                "responses": {
                    "200": {
                        "description": "Historique"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "💸 Retraits & Virements"
                ],
                "summary": "Demander un retrait Mobile Money",
                "description": "Conducteur uniquement. Soumet une demande qui sera traitée par un admin dans les 24 h.",
                "operationId": "withdrawalStore",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "amount",
                                    "provider",
                                    "phone_number"
                                ],
                                "properties": {
                                    "amount": {
                                        "description": "Montant XOF (min 500)",
                                        "type": "integer",
                                        "example": 15000,
                                        "minimum": 500
                                    },
                                    "provider": {
                                        "type": "string",
                                        "example": "mtn",
                                        "enum": [
                                            "mtn",
                                            "moov",
                                            "celtiis"
                                        ]
                                    },
                                    "phone_number": {
                                        "type": "string",
                                        "example": "97000000"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Demande enregistrée"
                    },
                    "403": {
                        "description": "Réservé aux conducteurs"
                    },
                    "422": {
                        "description": "Données invalides"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/withdrawals": {
            "get": {
                "tags": [
                    "💸 Retraits & Virements"
                ],
                "summary": "[ADMIN] Demandes de retrait",
                "operationId": "adminWithdrawals",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "enum": [
                                "pending",
                                "approved",
                                "rejected",
                                "failed"
                            ]
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Liste"
                    },
                    "403": {
                        "description": "Accès refusé"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/withdrawals/balance": {
            "get": {
                "tags": [
                    "💸 Retraits & Virements"
                ],
                "summary": "[ADMIN] Solde FedaPay disponible par opérateur",
                "operationId": "fedapayBalance",
                "responses": {
                    "200": {
                        "description": "Soldes"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "502": {
                        "description": "Erreur API FedaPay"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/withdrawals/{id}/process": {
            "post": {
                "tags": [
                    "💸 Retraits & Virements"
                ],
                "summary": "[ADMIN] Approuver & virer via FedaPay Payout",
                "description": "Vérifie le solde FedaPay, crée un Payout et envoie les fonds au conducteur. Notification push envoyée au conducteur.",
                "operationId": "withdrawalProcess",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Virement envoyé"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Demande introuvable"
                    },
                    "422": {
                        "description": "Déjà traitée / solde insuffisant"
                    },
                    "502": {
                        "description": "Échec FedaPay Payout"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/api/admin/withdrawals/{id}/reject": {
            "post": {
                "tags": [
                    "💸 Retraits & Virements"
                ],
                "summary": "[ADMIN] Rejeter une demande de retrait",
                "operationId": "withdrawalReject",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "reason"
                                ],
                                "properties": {
                                    "reason": {
                                        "type": "string",
                                        "example": "Solde insuffisant sur le compte MINIZON."
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Demande rejetée"
                    },
                    "403": {
                        "description": "Accès refusé"
                    },
                    "404": {
                        "description": "Demande introuvable"
                    },
                    "422": {
                        "description": "Déjà traitée ou motif manquant"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        }
    },
    "components": {
...
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "Sanctum",
                "scheme": "bearer"
            }
        }
    },
...
}